Split the nearby session code out of NearbyPlay and AppRuntime
No behaviour change. NearbyPlay.cpp was 680 lines and AppRuntime.cpp 701,
both past the size the modularity rule allows, and the next change touches
the session half of both -- so it moves first, on its own.

- NearbySession.cpp takes NearbyPlay's two-player half verbatim: seats,
  invitations, turns and the reserved ending. NearbyPlay.cpp keeps the score
  exchange and the notifications.
- The peer table both halves read moves from an anonymous namespace into
  NearbyPlay::detail in NearbyPlayState.h, which nothing else may include.
- AppRuntimeNearby.cpp takes BrainoApp's nearby* forwarders.

NearbyPlay.cpp is now 501 lines and AppRuntime.cpp 662. The flash figure
moves by 32 bytes, all of it the branch name in the build stamp.
