+++ CLAUDE.md
@@ -556,7 +556,7 @@ The same reasoning applies to any lock PlatformIO itself leaves in `~/.platformi
 
 ### Shared budgets
 
-Flash is global and nearly the binding constraint (2,462,145 / 3,145,728 bytes,
+Flash is global and nearly the binding constraint (2,462,177 / 3,145,728 bytes,
 **78.3%**; NimBLE plus the BT controller account for ~192 KB of that). RAM sits
 at 76,868 / 327,680 (23.5%) -- higher than it was, deliberately: RowList traded
 864 bytes of static RAM for zero heap traffic and storage diagnostics keep their
